storygen/requirements.txt
psycopg2-binary==2.9.9
openai==0.28.1
python-dotenv==1.0.0